refactor(export): move exceljs to an opt-in works-calendar/xlsx subpath
exceljs was the biggest thing breaking embedders. Even though it was
declared external and dynamically imported, the compiled core bundle
still contained a bare `import('exceljs')`, so any consumer bundler
tried to resolve it at build time and failed unless the consumer
happened to install exceljs (which pulls in unzipper -> optional
@aws-sdk/* Node deps that don't belong in a browser bundle).

Split the exporter:
- src/export/csvExport.ts — dependency-free CSV export + the shared
  row-shaping / sanitization / download helpers. Lives in the core.
- src/export/excelExport.ts — the exceljs path, now the only module
  that references exceljs. Reuses the CSV helpers and keeps the
  existing CSV fallback if exceljs isn't present at runtime.
- The built-in toolbar export (exportVisibleEvents) now produces CSV,
  so the core graph has zero exceljs references.
- index.ts exports `exportToCsv` from core; `exportToExcel` moves to
  the new `works-calendar/xlsx` subpath (second Vite lib entry ->
  dist/xlsx.js + dist/xlsx.d.ts, added to package.json exports).
  exceljs was already an optional peerDependency.

Verified: dist core + all chunks contain no `exceljs` reference (only
dist/xlsx.js does). The reference host now builds with no webpack
externals/fallbacks — removed the @aws-sdk/client-s3 shim that was
papering over this. Library type-check + 2832 tests green.

This is a breaking change for anyone importing `exportToExcel` from
the package root — they switch to `works-calendar/xlsx` and install
exceljs. Acceptable: the current embed surface doesn't really work, so
there are no such consumers to protect.
